+++ app/controllers/questions_controller.rb
@@ -1,4 +1,6 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
   def index
     @questions = Question.all
   end
@@ -13,10 +15,10 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -31,7 +33,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    question.destroy if current_user.author_of?(question)
     redirect_to questions_path
   end
 
@@ -41,7 +43,13 @@ class QuestionsController < ApplicationController
     @question ||= params[:id] ? Question.find(params[:id]) : Question.new
   end
 
+  def answer
+    @answer = question.answers.new
+  end
+
   helper_method :question
+
+  helper_method :answer
 
   def question_params
     params.require(:question).permit(:title, :body)
